AgglomertiveClustering
This algorithm uses dendrogram to find the optimal no of  clusters. Dendrogram is built by computing the euclidean distance between the clusters and combining those clusters that are the closest.
It starts forming the clusters by initializing considering all points as a cluster and then building a tree by computing the distances.
The part of the tree that has longest length is selected to find the no of clusters which is eventually considered as the optimal number.

Based on the number of clusters formed, the agglomertive clustering algorithm is used to find the points that lie in each clusters.